Do not restart pods twice
We originally put the `rollout restart` command in because in some
instances the `apply` did not restart the pods. This was in instances
when the commit sha that is being deployed has not changed, i.e when a
pipeline is being re run without any code changes.

Having the `rollout restart` run every time meant that we were
restarting the pods twice during each deployment.

Now we check to see if the last commit sha has changed and only call
`rollout restart` if it remains the same.