Retrieve CirclCI API token from K8S secret
We cannot find how to access the CircleCI API token except by doing
something very naughty.

Since not everyone is comfortable being naughty let's use a K8S secret
that can be accessed by the fb-deploy scripts in order to get the
CIRCLE_TOKEN instead of having to set it as an environment variable
in every pipeline.

As we are going to be setting the K8S context for the formbuilder-repos
namespace twice, pull that out into its' own script and source it when
required.