Use deployment name instead of image name
Some of our apps have inconsistencies between the name of the images
that are deployed and the name of the deployments within kubernetes
config.

Instead of taking the name of the image, get the name of the deployment
and match using that so that rollout, restart and status can be used
with the correct parameter.