Do not build images if they already exist
This is potentially a controversial PR.

For each ECR repo we are always building and pushing images even if
they have already been built and push. We can speed up two different
parts of the deployment if we check for an images existence first:

1. When deploying to Live. The same image tagged with the build SHA
   is used in both Test and Live.
2. Whenever we re run a deployment pipeline without any code changes.

Using `aws ecr describe-images` will look up the metadata for that
particular image. If it doesn't exist then returns an error message
and a non zero exit code (255 I think). If that happens then we know
to build the image. However if the command actually returns some
metadata then it's already been built before so we can skip building
and pushing and allow the deploy step to restart the pods which will
grab the already existing image with the existing build SHA.

Doing the above means that we can knock off a few minutes (between 2
and 4 for the submitter) for first time deployments of a code change,
and then for re runs it saves a lot more.

Finally we also move the authentication with the apps ECR repo to later
in the loop as we only want to do this if an actual Dockerfile exists.