Do not skip the build and push stage for live environments
The skip mechanism was there in order to stop building the image twice
if the SHA is the same. This is because test-dev, test-production use
the same container just with different tags: latest-test and
latest-live.

However the skip is checking whether the SHA exists, which it will do
because it just built it for latest-test and then skipping tagging for
latest-live.

This will unfortunately mean we will always be building the image twice,
at least for now.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>