Deploy testable branches for the editor app
This creates the ability to trigger deployment of specific branches for
the editor app for any github branch that has been prepended with
'testabe-'.

This means that developers can deploy out specific changes for the
editor and they can be tested independently of each other.

This process relies very heavily on two environment variables being set:

- editor_host
- app_name

The editor_host is set as a secret in the fb-editor-deploy repo, with
different values for Test and Live. This needs to be overridden in any
deployable branch to include the branch name. It is the thing that makes
the authentication mechanism work correctly. The editor app is the only
app that makes use of this environment variable presently. All other
apps will ignore the value set when a standard deployment is triggered.

The app_name is used to set the hostname for the app which Kubernetes
will use to create the necessary ingress rules as well as deployment
configuration.