Do not build latest tag unless runner app
We change the build script so that it will only tag an image with
'latest-test' or 'latest-live' if the app being deployed is one of the
runner apps, i.e fb-runner or fb-runner-node. Both of those apps do not
make use of the SHA as they are 'published' by the editor and publisher
apps respectively. The editor and the publisher are not able to see the
SHA which relatest the latest git commit as these are only available in
the CircleCI pipeline. As a result of this the editor and publisher need
to use an image which has been built and tagged with 'latest-test' or
'latest-live'.

If the app being deployed is not a runner type then the SHA relating to
the triggering git commit will be used as the image tag.