Correctly handle the AWS CLI response from ECR
Calling `aws cli describe-images` will return an error response if the
image does not exist. Because we set `-e` in the build script file this
means that the script will exit at that point.

What we want is to check whether the response returned a metadata
string, which it will if the image requested exists, or whether it
raised an error.

If the latter then the only way to avoid the script exiting is to use
`||` and run `build_and_push_sha_image` instead.