Build specific images base on application type
Some of our applications have two types. They are usually a web version
and a worker version, but sometimes they are an api type.

This change provides the ability to have the option for deployment
pipelines to be configured to build both the types in parallel. This is
done by setting an environment variable called IMAGE_TYPE to whatever
type you want to build in a given step.

If no IMAGE_TYPE is set then the script will build the application in
the normal way.