Handle unbound variable correctly
The IMAGE_TYPE variable may or may not be set. It depends on which app
is being deployed.

In order to avoid the unbound variable error being thrown as a result of
the `-u` setting in the script we have to add a `-` when checking the
variable exists. Bash will then be able to handle if the variable has
not been set to a value.