Target specific runner apps when restarting
Adjust the restart script so that it looks for apps specific to the type
of runner that is being deployed.

Requires the environment variable RUNNER_TYPE to be set in the
environment to be either fb-runner or fb-runner-node.

When calling rollout restart kubectl will resturn a string of the
deployment it restarted along with a response keyword such as
'restarted'. We need to filter these out when checking on the rollout
status for a deployment. There may well be other keywords we should
filter out in the future.