Do not echo in functions
When echo'ing anything inside a function that is mean to return a result
from a bash command, all the things that are echo'd are also returned
inside the string.

If you want to make use of the return value from a function roOnly return
the result and never anything else.