Deploy secondary no branching editor
This adds the commands necessary for deploying out the no branching
version of the editor in the test environment.

We set the DISABLE_BRANCHING environment variable in the CircleCI step.
If that is present then set the app_name to be
`fb-editor-test-no-branching` and pass editor_host and app_name and
disable_branching flags to the helm command.