Add scripts to rotate EKS tokens per namespace
This adds to ruby scripts:

- eks_token.rb
- eks_bearer_token.rb

Both scripts will delete the old secrets when they are run. They require
prompting by the dev running the script.

eks_token.rb will get the EKS token itself out of the CircleCI secret
for a given namespace and print to screen along with the correct
variable name that needs to be set in the CircleCI pipeline
configuration.

`ruby eks_token.rb <namespace>`

eks_bearer_token.rb gets the specific app secret prefixed with
'formbuilder-<app_name>' and prints out the name of that secret and the
name of the environment variable that should be used in the CircleCI
pipeline configuration.

`ruby eks_bearer_token.rb <namespace> <app_name>`

Co-authored-by: Chris Pymm <chris.pymm@digital.justice.gov.uk>
Co-authored-by: Matt Tei <matt.tei@digital.justice.gov.uk>
Co-authored-by: Hellema Ibrahim <hellema.ibrahim@digital.justice.gov.uk>
Co-authored-by: Steven Leighton <steven.leighton@digital.justice.gov.uk>